+++ System/Remote/Counter.hs
@@ -1,28 +1,18 @@
-{-# LANGUAGE BangPatterns #-}
+{-# OPTIONS_HADDOCK not-home #-}
+
 -- | This module defines a type for mutable, integer-valued counters.
 -- Counters are non-negative, monotonically increasing values and can
 -- be used to track e.g. the number of requests served since program
 -- start.  All operations on counters are thread-safe.
+--
+-- N.B. This module exists to maintain backwards compatibility with
+-- older versions of this library. New code should use the
+-- @System.Metrics.Counter@ module from the ekg-core package instead.
 module System.Remote.Counter
     (
-      Counter
-    , inc
-    , add
+      Counter.Counter
+    , Counter.inc
+    , Counter.add
     ) where
 
-import Data.IORef (atomicModifyIORef)
-import Prelude hiding (subtract)
-
-import System.Remote.Counter.Internal
-
--- | Increase the counter by one.
-inc :: Counter -> IO ()
-inc (C ref) = do
-    !_ <- atomicModifyIORef ref $ \ n -> let n' = n + 1 in (n', n')
-    return ()
-
--- | Increase the counter by the given amount.
-add :: Counter -> Int -> IO ()
-add (C ref) i = do
-    !_ <- atomicModifyIORef ref $ \ n -> let n' = n + i in (n', n')
-    return ()
+import qualified System.Metrics.Counter as Counter