+++ System/Remote/Monitoring.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE CPP #-}
+{-# LANGUAGE CPP, OverloadedStrings #-}
 
 -- | This module provides remote monitoring of a running process over
 -- HTTP.  It can be used to run an HTTP server that provides both a
@@ -31,30 +31,48 @@ module System.Remote.Monitoring
       -- * The monitoring server
       Server
     , serverThreadId
+    , serverMetricStore
     , forkServer
+    , forkServerNoHostname
+    , forkServerWith
+    , forkServerNoHostnameWith
 
-      -- * User-defined counters, gauges, and labels
+      -- * Defining metrics
       -- $userdefined
     , getCounter
     , getGauge
     , getLabel
+    , getDistribution
     ) where
 
-import Control.Concurrent (ThreadId, forkIO)
+import Control.Concurrent (ThreadId, myThreadId, throwTo)
+import Control.Exception (AsyncException(ThreadKilled), fromException)
 import qualified Data.ByteString as S
-import qualified Data.HashMap.Strict as M
-import Data.IORef (newIORef)
+import Data.Int (Int64)
+import qualified Data.Text as T
+import Data.Time.Clock.POSIX (getPOSIXTime)
 import Prelude hiding (read)
 
-import System.Remote.Common
-
+import qualified System.Metrics as Metrics
+import qualified System.Metrics.Counter as Counter
+import qualified System.Metrics.Distribution as Distribution
+import qualified System.Metrics.Gauge as Gauge
+import qualified System.Metrics.Label as Label
 import System.Remote.Snap
+import Network.Socket (withSocketsDo)
+
+#if __GLASGOW_HASKELL__ >= 706
+import Control.Concurrent (forkFinally)
+#else
+import Control.Concurrent (forkIO)
+import Control.Exception (SomeException, mask, try)
+#endif
 
 -- $configuration
 --
--- To use this module you must first enable GC statistics collection
--- in the run-time system.  To enable GC statistics collection, either
--- run your program with
+-- To make full use out of this module you must first enable GC
+-- statistics collection in the run-time system. To enable GC
+-- statistics collection, either run your program with
 --
 -- > +RTS -T
 --
@@ -73,118 +91,79 @@ import System.Remote.Snap
 
 -- $api
 -- To use the machine-readable REST API, send an HTTP GET request to
--- the host and port passed to 'forkServer'.  The following resources
--- (i.e. URLs) are available:
+-- the host and port passed to 'forkServer'.
 --
--- [\/] JSON object containing all counters, gauges and labels.
--- Counters, gauges, and labels are stored as nested objects under the
--- @counters@, @gauges@, and @labels@ attributes, respectively.
--- Content types: \"text\/html\" (default), \"application\/json\"
+-- The API is versioned to allow for API evolution. This document is
+-- for version 1. To ensure you're using this version, append @?v=1@
+-- to your resource URLs. Omitting the version number will give you
+-- the latest version of the API.
 --
--- [\/combined] Flattened JSON object containing all counters, gauges,
--- and labels.  Content types: \"application\/json\"
+-- The following resources (i.e. URLs) are available:
 --
--- [\/counters] JSON object containing all counters.  Content types:
+-- [\/] JSON object containing all metrics. Metrics are stored as
+-- nested objects, with one new object layer per \".\" in the metric
+-- name (see example below.) Content types: \"text\/html\" (default),
 -- \"application\/json\"
 --
--- [\/counters/\<counter name\>] Value of a single counter, as a
--- string.  The name should be UTF-8 encoded.  Content types:
--- \"text\/plain\"
+-- [\/\<namespace\>/\<metric\>] JSON object for a single metric. The
+-- metric name is created by converting all \"\/\" to \".\". Example:
+-- \"\/foo\/bar\" corresponds to the metric \"foo.bar\". Content
+-- types: \"application\/json\"
 --
--- [\/gauges] JSON object containing all gauges.  Content types:
--- \"application\/json\"
+-- Each metric is returned as an object containing a @type@ field.  Available types
+-- are:
 --
--- [\/gauges/\<gauge name\>] Value of a single gauge, as a string.
--- The name should be UTF-8 encoded.  Content types: \"text\/plain\"
+--  * \"c\" - 'Counter.Counter'
 --
--- [\/labels] JSON object containing all labels.  Content types:
--- \"application\/json\"
+--  * \"g\" - 'Gauge.Gauge'
 --
--- [\/labels/\<label name\>] Value of a single label, as a string.
--- The name should be UTF-8 encoded.  Content types: \"text\/plain\"
+--  * \"l\" - 'Label.Label'
 --
--- Counters, gauges and labels are stored as attributes of the
--- returned JSON objects, one attribute per counter, gauge or label.
--- In addition to user-defined counters, gauges, and labels, the below
--- built-in counters and gauges are also returned.  Furthermore, the
--- top-level JSON object of any resource contains the
--- @server_timestamp_millis@ attribute, which indicates the server
--- time, in milliseconds, when the sample was taken.
+--  * \"d\" - 'Distribution.Distribution'
 --
--- Built-in counters:
+-- In addition to the @type@ field, there are metric specific fields:
 --
--- [@bytes_allocated@] Total number of bytes allocated
+--  * Counters, gauges, and labels: the @val@ field contains the
+--    actual value (i.e. an integer or a string).
 --
--- [@num_gcs@] Number of garbage collections performed
+--  * Distributions: the @mean@, @variance@, @count@, @sum@, @min@,
+--    and @max@ fields contain their statistical equivalents.
 --
--- [@num_bytes_usage_samples@] Number of byte usage samples taken
+-- Example of a response containing the metrics \"myapp.visitors\" and
+-- \"myapp.args\":
 --
--- [@cumulative_bytes_used@] Sum of all byte usage samples, can be
--- used with @numByteUsageSamples@ to calculate averages with
--- arbitrary weighting (if you are sampling this record multiple
--- times).
---
--- [@bytes_copied@] Number of bytes copied during GC
---
--- [@mutator_cpu_seconds@] CPU time spent running mutator threads.
--- This does not include any profiling overhead or initialization.
---
--- [@mutator_wall_seconds@] Wall clock time spent running mutator
--- threads.  This does not include initialization.
---
--- [@gc_cpu_seconds@] CPU time spent running GC
---
--- [@gc_wall_seconds@] Wall clock time spent running GC
---
--- [@cpu_seconds@] Total CPU time elapsed since program start
---
--- [@wall_seconds@] Total wall clock time elapsed since start
---
--- Built-in gauges:
---
--- [@max_bytes_used@] Maximum number of live bytes seen so far
---
--- [@current_bytes_used@] Current number of live bytes
---
--- [@current_bytes_slop@] Current number of bytes lost to slop
---
--- [@max_bytes_slop@] Maximum number of bytes lost to slop at any one time so far
---
--- [@peak_megabytes_allocated@] Maximum number of megabytes allocated
---
--- [@par_tot_bytes_copied@] Number of bytes copied during GC, minus
--- space held by mutable lists held by the capabilities.  Can be used
--- with 'parMaxBytesCopied' to determine how well parallel GC utilized
--- all cores.
---
--- [@par_avg_bytes_copied@] Deprecated alias for
--- @par_tot_bytes_copied@.
---
--- [@par_max_bytes_copied@] Sum of number of bytes copied each GC by
--- the most active GC thread each GC. The ratio of
--- @par_tot_bytes_copied@ divided by @par_max_bytes_copied@ approaches
--- 1 for a maximally sequential run and approaches the number of
--- threads (set by the RTS flag @-N@) for a maximally parallel run.
+-- > {
+-- >   "myapp": {
+-- >     "visitors": {
+-- >       "val": 10,
+-- >       "type": "c"
+-- >     },
+-- >     "args": {
+-- >       "val": "--a-flag",
+-- >       "type": "l"
+-- >     }
+-- >   }
+-- > }
 
 -- $userdefined
--- The monitoring server can store and serve user-defined,
--- integer-valued counters and gauges, and string-valued labels.  A
+-- The monitoring server can store and serve integer-valued counters
+-- and gauges, string-valued labels, and statistical distributions. A
 -- counter is a monotonically increasing value (e.g. TCP connections
--- established since program start.) A gauge is a variable value
--- (e.g. the current number of concurrent connections.) A label is a
+-- established since program start.) A gauge is a variable value (e.g.
+-- the current number of concurrent connections.) A label is a
 -- free-form string value (e.g. exporting the command line arguments
--- or host name.)  Each counter, gauge, and label is associated with a
--- name, which is used when it is displayed in the UI or returned in a
--- JSON object.
+-- or host name.) A distribution is a statistic summary of events
+-- (e.g. processing time per request.) Each metric is associated with
+-- a name, which is used when it is displayed in the UI or returned in
+-- a JSON object.
 --
--- Even though it's technically possible to have a counter and a gauge
--- with the same name, associated with the same server, it's not
--- recommended as it might make it harder for clients to distinguish
--- the two.
+-- Metrics share the same namespace so it's not possible to create
+-- e.g. a counter and a gauge with the same. Attempting to do so will
+-- result in an 'error'.
 --
 -- To create and use a counter, simply call 'getCounter' to create it
--- and then call e.g. 'System.Remote.Counter.inc' or
--- 'System.Remote.Counter.add' to modify its value.  Example:
+-- and then call e.g. 'Counter.inc' or 'Counter.add' to modify its
+-- value. Example:
 --
 -- > main = do
 -- >     handle <- forkServer "localhost" 8000
@@ -195,31 +174,148 @@ import System.Remote.Snap
 -- >     loop
 --
 -- To create a gauge, use 'getGauge' instead of 'getCounter' and then
--- call e.g. 'System.Remote.Gauge.set' or
--- 'System.Remote.Gauge.modify'.  Similar for labels.
+-- call e.g. 'System.Remote.Gauge.set'. Similar for the other metric
+-- types.
+--
+-- It's also possible to register metrics directly using the
+-- @System.Metrics@ module in the ekg-core package. This gives you a
+-- bit more control over how metric values are retrieved.
 
 ------------------------------------------------------------------------
 -- * The monitoring server
 
--- | The thread ID of the server.  You can kill the server by killing
--- this thread (i.e. by throwing it an asynchronous exception.)
-serverThreadId :: Server -> ThreadId
-serverThreadId = threadId
+-- | A handle that can be used to control the monitoring server.
+-- Created by 'forkServer'.
+data Server = Server {
+      -- | The thread ID of the server. You can kill the server by
+      -- killing this thread (i.e. by throwing it an asynchronous
+      -- exception.)
+      serverThreadId :: {-# UNPACK #-} !ThreadId
+
+      -- | The metric store associated with the server. If you want to
+      -- add metric to the default store created by 'forkServer' you
+      -- need to use this function to retrieve it.
+    , serverMetricStore :: {-# UNPACK #-} !Metrics.Store
+    }
+
+-- | Like 'forkServerWith', but creates a default metric store with
+-- some predefined metrics. The predefined metrics are those given in
+-- 'System.Metrics.registerGcMetrics'.
+forkServer :: S.ByteString  -- ^ Host to listen on (e.g. \"localhost\")
+           -> Int           -- ^ Port to listen on (e.g. 8000)
+           -> IO Server
+forkServer host port = do
+    store <- Metrics.newStore
+    Metrics.registerGcMetrics store
+    forkServerMaybeHostnameWith store (Just host) port
+
+-- | Create a server with prefined metrics from
+-- 'System.Metrics.registerGcMetrics', listening on all interfaces.
+-- If you are running EKG on a private network (including virtual
+-- private network), it may be appropriate to bind to all interfaces,
+-- not only localhost.
+forkServerNoHostname :: Int           -- ^ Port to listen on (e.g. 8000)
+                     -> IO Server
+forkServerNoHostname port = do
+    store <- Metrics.newStore
+    Metrics.registerGcMetrics store
+    forkServerMaybeHostnameWith store Nothing port
 
 -- | Start an HTTP server in a new thread.  The server replies to GET
 -- requests to the given host and port.  The host argument can be
 -- either a numeric network address (dotted quad for IPv4,
 -- colon-separated hex for IPv6) or a hostname (e.g. \"localhost\".)
 -- The client can control the Content-Type used in responses by
--- setting the Accept header.  At the moment three content types are
--- available: \"application\/json\", \"text\/html\", and
--- \"text\/plain\".
-forkServer :: S.ByteString  -- ^ Host to listen on (e.g. \"localhost\")
-           -> Int           -- ^ Port to listen on (e.g. 8000)
-           -> IO Server
-forkServer host port = do
-    counters <- newIORef M.empty
-    gauges <- newIORef M.empty
-    labels <- newIORef M.empty
-    tid <- forkIO $ startServer counters gauges labels host port
-    return $! Server tid counters gauges labels
+-- setting the Accept header.  At the moment two content types are
+-- available: \"application\/json\" and \"text\/html\".
+--
+-- Registers the following counter, used by the UI:
+--
+-- [@ekg.server_time_ms@] The server time when the sample was taken,
+-- in milliseconds.
+--
+-- Note that this function, unlike 'forkServer', doesn't register any
+-- other predefined metrics. This allows other libraries to create and
+-- provide a metric store for use with this library. If the metric
+-- store isn't created by you and the creator doesn't register the
+-- metrics registered by 'forkServer', you might want to register them
+-- yourself.
+forkServerWith :: Metrics.Store  -- ^ Metric store
+               -> S.ByteString   -- ^ Host to listen on (e.g. \"localhost\")
+               -> Int            -- ^ Port to listen on (e.g. 8000)
+               -> IO Server
+forkServerWith store host port =
+    forkServerMaybeHostnameWith store (Just host) port
+
+-- | Start an HTTP server in a new thread, with the specified metrics
+-- store, listening on all interfaces.  Other than accepting requests
+-- to any hostname, this is the same as `forkServerWith`.
+forkServerNoHostnameWith :: Metrics.Store  -- ^ Metric store
+                         -> Int            -- ^ Port to listen on (e.g. 8000)
+                         -> IO Server
+forkServerNoHostnameWith store port =
+    forkServerMaybeHostnameWith store Nothing port
+
+forkServerMaybeHostnameWith :: Metrics.Store  -- ^ Metric store
+                            -> Maybe S.ByteString   -- ^ Host to listen on (e.g. \"localhost\")
+                            -> Int            -- ^ Port to listen on (e.g. 8000)
+                            -> IO Server
+forkServerMaybeHostnameWith store host port = do
+    Metrics.registerCounter "ekg.server_timestamp_ms" getTimeMs store
+    me <- myThreadId
+    tid <- withSocketsDo $ forkFinally (startServer store host port) $ \ r ->
+        case r of
+            Right _ -> return ()
+            Left e  -> case fromException e of
+                Just ThreadKilled -> return ()
+                _                 -> throwTo me e
+    return $! Server tid store
+  where
+    getTimeMs :: IO Int64
+    getTimeMs = (round . (* 1000)) `fmap` getPOSIXTime
+
+------------------------------------------------------------------------
+-- * Defining metrics
+
+-- | Return a new, zero-initialized counter associated with the given
+-- name and server. Multiple calls to 'getCounter' with the same
+-- arguments will result in an 'error'.
+getCounter :: T.Text  -- ^ Counter name
+           -> Server  -- ^ Server that will serve the counter
+           -> IO Counter.Counter
+getCounter name server = Metrics.createCounter name (serverMetricStore server)
+
+-- | Return a new, zero-initialized gauge associated with the given
+-- name and server. Multiple calls to 'getGauge' with the same
+-- arguments will result in an 'error'.
+getGauge :: T.Text  -- ^ Gauge name
+         -> Server  -- ^ Server that will serve the gauge
+         -> IO Gauge.Gauge
+getGauge name server = Metrics.createGauge name (serverMetricStore server)
+
+-- | Return a new, empty label associated with the given name and
+-- server. Multiple calls to 'getLabel' with the same arguments will
+-- result in an 'error'.
+getLabel :: T.Text  -- ^ Label name
+         -> Server  -- ^ Server that will serve the label
+         -> IO Label.Label
+getLabel name server = Metrics.createLabel name (serverMetricStore server)
+
+-- | Return a new distribution associated with the given name and
+-- server. Multiple calls to 'getDistribution' with the same arguments
+-- will result in an 'error'.
+getDistribution :: T.Text  -- ^ Distribution name
+                -> Server  -- ^ Server that will serve the distribution
+                -> IO Distribution.Distribution
+getDistribution name server =
+    Metrics.createDistribution name (serverMetricStore server)
+
+------------------------------------------------------------------------
+-- Backwards compatibility shims
+
+#if __GLASGOW_HASKELL__ < 706
+forkFinally :: IO a -> (Either SomeException a -> IO ()) -> IO ThreadId
+forkFinally action and_then =
+  mask $ \restore ->
+    forkIO $ try (restore action) >>= and_then
+#endif