+++ System/Remote/Snap.hs
@@ -7,31 +7,30 @@ module System.Remote.Snap
 
 import Control.Applicative ((<$>), (<|>))
 import Control.Exception (throwIO)
-import Control.Monad (join, unless)
 import Control.Monad.IO.Class (liftIO)
-import qualified Data.Aeson.Types as A
 import qualified Data.ByteString as S
 import qualified Data.ByteString.Char8 as S8
+import Data.Maybe (fromMaybe)
+import Data.Function (on)
 import qualified Data.HashMap.Strict as M
-import Data.IORef (IORef)
 import qualified Data.List as List
-import qualified Data.Map as Map
-import Data.Maybe (fromMaybe, listToMaybe)
-import qualified Data.Text as T
 import qualified Data.Text.Encoding as T
+import Data.Word (Word8)
 import Network.Socket (NameInfoFlag(NI_NUMERICHOST), addrAddress, getAddrInfo,
                        getNameInfo)
 import Prelude hiding (read)
-import Snap.Core (MonadSnap, Request, Snap, finishWith, getHeaders, getRequest,
-                  getResponse, method, Method(GET), modifyResponse, pass, route,
-                  rqURI, rqParams, rqPathInfo, setContentType, setResponseStatus,
-                  writeBS, writeLBS, setResponseCode, setContentLength, )
+import Snap.Core (MonadSnap, Request, Snap, finishWith, getHeader, getRequest,
+                  getResponse, method, Method(GET), modifyResponse, pass,
+                  rqURI, rqPathInfo,
+                  setContentType, setContentLength, setResponseStatus, setResponseCode,
+                  writeBS, writeLBS)
 import Snap.Http.Server (httpServe)
 import qualified Snap.Http.Server.Config as Config
 import Snap.Util.FileServe (defaultMimeTypes)
 import System.FilePath (takeExtension)
 
-import System.Remote.Common
+import System.Metrics
+import System.Remote.Json
 import Data.FileEmbed (embedDir)
 
 
@@ -53,51 +52,40 @@ getNumericHostAddress host = do
     unsupportedAddressError = throwIO $
         userError $ "unsupported address: " ++ S8.unpack host
 
-startServer :: IORef Counters -> IORef Gauges -> IORef Labels
-            -> S.ByteString  -- ^ Host to listen on (e.g. \"localhost\")
+startServer :: Store
+            -> Maybe S.ByteString  -- ^ Host to listen on (e.g. \"localhost\")
             -> Int           -- ^ Port to listen on (e.g. 8000)
             -> IO ()
-startServer counters gauges labels host port = do
+startServer store m_host port = do
     -- Snap doesn't allow for non-numeric host names in
     -- 'Snap.setBind'. We work around that limitation by converting a
     -- possible non-numeric host name to a numeric address.
-    numericHost <- getNumericHostAddress host
+    setBind <- case m_host of
+        Just host -> do
+            numericHost <- getNumericHostAddress host
+            return $ Config.setHostname host . Config.setBind numericHost
+        Nothing -> return id
     let conf = Config.setVerbose False $
                Config.setErrorLog Config.ConfigNoLog $
                Config.setAccessLog Config.ConfigNoLog $
                Config.setPort port $
-               Config.setHostname host $
-               Config.setBind numericHost $
+               setBind $
                Config.defaultConfig
-    httpServe conf (monitor counters gauges labels)
+    httpServe conf (monitor store)
 
 -- | A handler that can be installed into an existing Snap application.
-monitor :: IORef Counters -> IORef Gauges -> IORef Labels -> Snap ()
-monitor counters gauges labels = do
-    route [
-          ("", method GET (format "application/json"
-                           (serveAll counters gauges labels)))
-        , ("combined", method GET (format "application/json"
-                                   (serveCombined counters gauges labels)))
-        , ("counters", method GET (format "application/json"
-                                   (serveMany counters)))
-        , ("counters/:name", method GET (format "text/plain"
-                                         (serveOne counters)))
-        , ("gauges", method GET (format "application/json"
-                                 (serveMany gauges)))
-        , ("gauges/:name", method GET (format "text/plain"
-                                       (serveOne gauges)))
-        , ("labels", method GET (format "application/json"
-                                 (serveMany labels)))
-        , ("labels/:name", method GET (format "text/plain"
-                                       (serveOne labels)))
-        ]
+monitor :: Store -> Snap ()
+monitor store = do
+    (jsonHandler $ serve store)
         <|>
         serveAssets
+  where
+    jsonHandler = wrapHandler "application/json"
+    wrapHandler fmt handler = method GET $ format fmt $ handler
 
 -- | The Accept header of the request.
 acceptHeader :: Request -> Maybe S.ByteString
-acceptHeader req = S.intercalate "," <$> getHeaders "Accept" req
+acceptHeader req = getHeader "Accept" req
 
 -- | Runs a Snap monad action only if the request's Accept header
 -- matches the given MIME type.
@@ -109,53 +97,58 @@ format fmt action = do
         Just hdr | hdr == fmt -> action
         _ -> pass
 
--- | Serve a collection of counters or gauges, as a JSON object.
-serveMany :: (Ref r t, A.ToJSON t) => IORef (M.HashMap T.Text r) -> Snap ()
-serveMany mapRef = do
-    modifyResponse $ setContentType "application/json"
-    bs <- liftIO $ buildMany mapRef
-    writeLBS bs
-{-# INLINABLE serveMany #-}
-
 -- | Serve all counter, gauges and labels, built-in or not, as a
 -- nested JSON object.
-serveAll :: IORef Counters -> IORef Gauges -> IORef Labels -> Snap ()
-serveAll counters gauges labels = do
+serve :: MonadSnap m => Store -> m ()
+serve store = do
     req <- getRequest
-    -- Workaround: Snap still matches requests to /foo to this handler
-    -- if the Accept header is "application/json", even though such
-    -- requests ought to go to the 'serveOne' handler.
-    unless (S.null $ rqPathInfo req) pass
     modifyResponse $ setContentType "application/json"
-    bs <- liftIO $ buildAll counters gauges labels
-    writeLBS bs
+    if S.null (rqPathInfo req)
+        then serveAll
+        else serveOne (rqPathInfo req)
+  where
+    serveAll = do
+        metrics <- liftIO $ sampleAll store
+        writeLBS $ encodeAll metrics
+    serveOne pathInfo = do
+        let segments  = S8.split '/' pathInfo
+            nameBytes = S8.intercalate "." segments
+        case T.decodeUtf8' nameBytes of
+            Left _ -> do
+                modifyResponse $ setResponseStatus 400 "Bad Request"
+                r <- getResponse
+                finishWith r
+            Right name -> do
+                metrics <- liftIO $ sampleAll store
+                case M.lookup name metrics of
+                    Nothing -> pass
+                    Just metric -> writeLBS $ encodeOne metric
 
--- | Serve all counters and gauges, built-in or not, as a flattened
--- JSON object.
-serveCombined :: IORef Counters -> IORef Gauges -> IORef Labels -> Snap ()
-serveCombined counters gauges labels = do
-    modifyResponse $ setContentType "application/json"
-    bs <- liftIO $ buildCombined counters gauges labels
-    writeLBS bs
+------------------------------------------------------------------------
+-- Utilities for working with accept headers
 
--- | Serve a single counter, as plain text.
-serveOne :: (Ref r t, Show t) => IORef (M.HashMap T.Text r) -> Snap ()
-serveOne refs = do
-    modifyResponse $ setContentType "text/plain"
-    req <- getRequest
-    let mname = T.decodeUtf8 <$> join
-                (listToMaybe <$> Map.lookup "name" (rqParams req))
-    case mname of
-        Nothing -> pass
-        Just name -> do
-            mbs <- liftIO $ buildOne refs name
-            case mbs of
-                Just bs -> writeBS bs
-                Nothing -> do
-                    modifyResponse $ setResponseStatus 404 "Not Found"
-                    r <- getResponse
-                    finishWith r
-{-# INLINABLE serveOne #-}
+-- | Parse the HTTP accept string to determine supported content types.
+parseHttpAccept :: S.ByteString -> [S.ByteString]
+parseHttpAccept = List.map fst
+                . List.sortBy (rcompare `on` snd)
+                . List.map grabQ
+                . S.split 44 -- comma
+  where
+    rcompare :: Double -> Double -> Ordering
+    rcompare = flip compare
+    grabQ s =
+        let (s', q) = breakDiscard 59 s -- semicolon
+            (_, q') = breakDiscard 61 q -- equals sign
+         in (trimWhite s', readQ $ trimWhite q')
+    readQ s = case reads $ S8.unpack s of
+                (x, _):_ -> x
+                _ -> 1.0
+    trimWhite = S.dropWhile (== 32) -- space
+
+breakDiscard :: Word8 -> S.ByteString -> (S.ByteString, S.ByteString)
+breakDiscard w s =
+    let (x, y) = S.break (== w) s
+    in (x, S.drop 1 y)
 
 -- | Serve the embedded assets.
 serveAssets :: MonadSnap m => m ()